main: split string decoding routines into their own files.
closes #26.
